requirements.txt
matplotlib
tensorflow
notebook
pysoundfile
pillow
h5py
pandas
